Add harvester_submit_week_for_approval.py
The start of a script that will log into Harvest via the normal web
interface and submit the most recent full week's time sheet for
approval.

It uses Selenium and the Firefox WebDriver. It works with the
`-headless` option, but I've commented that out for testing. Right now,
it can open the Harvest sign-in page and log in.