Draft of submitting week for approval
First pass, doesn't work yet.

This will determine last Friday's date and visit the page corresponding
to that week. It will then try to submit that week's timesheet.

The `wait` on the `script` tag contents waiting for the "week" page to
load isn't working. Looks like you can't wait on the contents of a
`<script>` tag.

Moved the `wait` variable outside of `login` so it can be used in
`submit_week_for_approval`.