submit_week_for_approval: Fix page load waiter; Click confirm button
Instead of waiting on an "invisible" `<script>` tag that Selenium
doesn't know how to do, check that the "week" page is loaded a different
way: by checking that the year appears in the header and the day appears
in a column header.

Uncomment the rest of the function.

We weren't clicking the confirmation button (`.approval-confirmation`).
Added a `click()` call there.