submit_week_for_approval: Fix date page load check
This time the test for page load using the `friday` date didn't work.
Not sure why, it should be the same and the text and element looks
right. The only thing I can think of is that Selenium isn't smart enough
to know there are multiple elements with the same class name and so it
was comparing the wrong one, and last time when it worked it was just a
fluke that happened to work the first time.

Replace the CSS selector with an XPath selector that will get a "unique"
element for us and still check that the correct date appears (here,
formatted as "Wednesday, 24 January").