Check that there are time entries for every day in the week
Before submitting the time sheet for approval (and before doing anything
with Selenium), ensure that there's at least one time entry logged for
each day in the week.

To do this, we query the Harvest API with newly-required account ID and
API token values. This gives us a bunch of time entries. We check them
to see if they fell in the last week. If all weekdays from the last week
have at least one time entry, then the rest of the program executes and
we run Selenium to submit the time sheet.

If any day had missing entries, an exception is raised and the program
exits.